requirements-dev.txt
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
werkzeug==2.0.2
pytest==6.2.5
pytest-cov==3.0.0
flake8==3.9.2
psycopg2==2.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy-Utils==0.37.8
marshmallow-sqlalchemy==0.26.1
marshmallow==3.13.0
flask-marshmallow==0.14.0
Flask-JWT-Extended==4.3.0
Flask-Cors==3.0.10
typing_extensions==3.10.0.0
requests==2.26.0
pytz==2021.3
psutil==5.8.0
Faker==8.14.1
ipython==7.28.0